Reimplemented menu extras as inbox extensions
Removed SystemMenuExtras extension
Removed MenuExtra files in CairoDesktop
Moved ShellFolders extension to CairoExtensionSamples
