--- Cairo Desktop/Cairo Desktop/MenuExtraSearchResult.xaml
@@ -1,79 +1,0 @@
-﻿<UserControl x:Class="CairoDesktop.MenuExtraSearchResult"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:l10n="clr-namespace:CairoDesktop.Localization;assembly=CairoDesktop.Localization">
-    <MenuItem Click="MenuItem_OnClick"
-              HorizontalAlignment="Stretch"
-              VerticalAlignment="Stretch"
-              CommandParameter="{Binding}"
-              ToolTip="{Binding Path=PathDisplay}"
-              Header="{Binding}">
-        <MenuItem.HeaderTemplate>
-            <ItemContainerTemplate>
-                <StackPanel Orientation="Horizontal"
-                            Margin="0"
-                            HorizontalAlignment="Stretch">
-                    <Image
-                        Source="{Binding Path=Icon, Mode=OneWay, FallbackValue={StaticResource NullIcon}, TargetNullValue={StaticResource NullIcon}}"
-                        Height="32"
-                        Width="32"
-                        Margin="0,0,10,0" />
-                    <StackPanel Orientation="Vertical"
-                                VerticalAlignment="Center">
-                        <TextBlock Foreground="{DynamicResource ResourceKey=SearchResultText}"
-                                   FontSize="{DynamicResource ResourceKey=LargeFontSize}"
-                                   FontFamily="{DynamicResource ResourceKey=GlobalFontFamily}"
-                                   Text="{Binding Path=Name}"
-                                   TextTrimming="CharacterEllipsis"
-                                   HorizontalAlignment="Left"
-                                   Width="268" />
-                        <StackPanel Orientation="Horizontal"
-                                    MaxWidth="268">
-                            <TextBlock Foreground="{DynamicResource ResourceKey=SearchResultSubtext}"
-                                       FontSize="{DynamicResource ResourceKey=MediumFontSize}"
-                                       FontFamily="{DynamicResource ResourceKey=GlobalFontFamily}"
-                                       Text="{Binding Path=(l10n:DisplayString.sSearch_LastModified)}"
-                                       TextTrimming="CharacterEllipsis"
-                                       Margin="0,0,3,0"
-                                       HorizontalAlignment="Left" />
-                            <TextBlock Foreground="{DynamicResource ResourceKey=SearchResultSubtext}"
-                                       FontSize="{DynamicResource ResourceKey=MediumFontSize}"
-                                       FontFamily="{DynamicResource ResourceKey=GlobalFontFamily}"
-                                       Text="{Binding Path=DateModified}"
-                                       TextTrimming="CharacterEllipsis"
-                                       HorizontalAlignment="Left" />
-                        </StackPanel>
-                    </StackPanel>
-                </StackPanel>
-            </ItemContainerTemplate>
-        </MenuItem.HeaderTemplate>
-        <MenuItem.Style>
-            <Style TargetType="{x:Type MenuItem}">
-                <Setter Property="Template">
-                    <Setter.Value>
-                        <ControlTemplate TargetType="{x:Type MenuItem}">
-                            <Border BorderThickness="0"
-                                    Background="Transparent"
-                                    Name="Border">
-                                <ContentPresenter Margin="15,0"
-                                                  x:Name="HeaderHost"
-                                                  RecognizesAccessKey="True"
-                                                  ContentSource="Header" />
-                            </Border>
-                            <ControlTemplate.Triggers>
-                                <Trigger Property="IsMouseOver"
-                                         Value="True">
-                                    <Trigger.Setters>
-                                        <Setter Property="Background"
-                                                Value="{DynamicResource ResourceKey=SearchResultHover}"
-                                                TargetName="Border" />
-                                    </Trigger.Setters>
-                                </Trigger>
-                            </ControlTemplate.Triggers>
-                        </ControlTemplate>
-                    </Setter.Value>
-                </Setter>
-            </Style>
-        </MenuItem.Style>
-    </MenuItem>
-</UserControl>
