--- Cairo Desktop/Cairo Desktop/MenuExtraVolume.xaml
@@ -1,30 +1,0 @@
-﻿<UserControl x:Class="CairoDesktop.MenuExtraVolume"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:l10n="clr-namespace:CairoDesktop.Localization;assembly=CairoDesktop.Localization">
-    <Menu Style="{StaticResource CairoMenuBarMainContainerStyle}">
-        <MenuItem x:Name="miOpenVolume"
-                  Style="{StaticResource CairoMenuItemButtonMenuStyle}"
-                  ItemContainerStyle="{StaticResource CairoMenuItemContainerStyle}"
-                  Click="miOpenVolume_Click"
-                  ToolTip="{Binding Path=(l10n:DisplayString.sMenuBar_Volume)}"
-                  ToolTipService.Placement="Bottom">
-            <MenuItem.Header>
-                <Image Source="{DynamicResource VolumeIcon}"
-                       Width="38"
-                       Height="22"
-                       Name="imgOpenVolume" />
-            </MenuItem.Header>
-            <MenuItem.ContextMenu>
-                <ContextMenu>
-                    <MenuItem Header="{Binding Path=(l10n:DisplayString.sMenuBar_OpenSoundSettings)}"
-                              Click="miOpenSoundSettings_Click"
-                              Name="miOpenSoundSettings" />
-                    <MenuItem Header="{Binding Path=(l10n:DisplayString.sMenuBar_OpenVolumeMixer)}"
-                              Click="miOpenVolumeMixer_Click"
-                              Name="miOpenVolumeMixer" />
-                </ContextMenu>
-            </MenuItem.ContextMenu>
-        </MenuItem>
-    </Menu>
-</UserControl>
