+++ Cairo Desktop/CairoDesktop.MenuBarExtensions/Search.xaml
@@ -1,7 +1,7 @@
-﻿<UserControl x:Class="CairoDesktop.MenuExtraSearch"
+﻿<UserControl x:Class="CairoDesktop.MenuBarExtensions.Search"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:cairo="clr-namespace:CairoDesktop"
+             xmlns:controls="clr-namespace:CairoDesktop.MenuBarExtensions"
              xmlns:l10n="clr-namespace:CairoDesktop.Localization;assembly=CairoDesktop.Localization">
     <Menu Style="{StaticResource CairoMenuBarMainContainerStyle}">
         <MenuItem x:Name="CairoSearchMenu"
@@ -67,7 +67,7 @@
                                             <Border BorderBrush="Transparent"
                                                     BorderThickness="0"
                                                     Background="{TemplateBinding Background}">
-                                                <cairo:MenuExtraSearchResult DataContext="{Binding}" />
+                                                <controls:SearchResult DataContext="{Binding}" />
                                             </Border>
                                         </ControlTemplate>
                                     </Setter.Value>
