+++ Cairo Desktop/CairoDesktop.MenuBarExtensions/SearchResult.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="CairoDesktop.Extensions.SystemMenuExtras.SearchResult"
+﻿<UserControl x:Class="CairoDesktop.MenuBarExtensions.SearchResult"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:l10n="clr-namespace:CairoDesktop.Localization;assembly=CairoDesktop.Localization">
