+++ Cairo Desktop/CairoDesktop.MenuBarExtensions/Volume.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="CairoDesktop.Extensions.SystemMenuExtras.Volume"
+﻿<UserControl x:Class="CairoDesktop.MenuBarExtensions.Volume"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:l10n="clr-namespace:CairoDesktop.Localization;assembly=CairoDesktop.Localization">
