+++ Cairo Desktop/CairoDesktop.MenuBarExtensions/Volume.xaml.cs
@@ -5,7 +5,7 @@ using System.Windows.Media;
 using System.Windows.Threading;
 using ManagedShell.Common.Helpers;
 
-namespace CairoDesktop.Extensions.SystemMenuExtras
+namespace CairoDesktop.MenuBarExtensions
 {
     public partial class Volume : UserControl
     {
