--- Cairo Desktop/Extensions/CairoDesktop.Extensions.SystemMenuExtras/ActionCenter.xaml
@@ -1,22 +1,0 @@
-﻿<UserControl x:Class="CairoDesktop.Extensions.SystemMenuExtras.ActionCenter"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:l10n="clr-namespace:CairoDesktop.Localization;assembly=CairoDesktop.Localization">
-    <Menu Style="{StaticResource CairoMenuBarMainContainerStyle}">
-        <MenuItem x:Name="miOpenActionCenter"
-                  Style="{StaticResource CairoMenuItemButtonMenuStyle}"
-                  ItemContainerStyle="{StaticResource CairoMenuItemContainerStyle}"
-                  Click="miOpenActionCenter_Click"
-                  MouseEnter="miOpenActionCenter_MouseEnter"
-                  MouseLeave="miOpenActionCenter_MouseLeave"
-                  ToolTip="{Binding Path=(l10n:DisplayString.sMenuBar_OpenActionCenter)}"
-                  ToolTipService.Placement="Bottom">
-            <MenuItem.Header>
-                <Image Name="ActionCenterIcon"
-                       Source="{DynamicResource ActionCenterIcon}"
-                       Width="38"
-                       Height="22" />
-            </MenuItem.Header>
-        </MenuItem>
-    </Menu>
-</UserControl>
