--- Cairo Desktop/Extensions/CairoDesktop.Extensions.SystemMenuExtras/Clock.xaml
@@ -1,39 +1,0 @@
-﻿<UserControl x:Class="CairoDesktop.Extensions.SystemMenuExtras.Clock"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:l10n="clr-namespace:CairoDesktop.Localization;assembly=CairoDesktop.Localization">
-    <Menu Style="{StaticResource CairoMenuBarMainContainerStyle}">
-        <MenuItem x:Name="ClockMenuItem"
-                  ItemContainerStyle="{StaticResource CairoMenuItemContainerStyle}"
-                  SubmenuOpened="ClockMenuItem_SubmenuOpened">
-            <MenuItem.Header>
-                <TextBlock Name="dateText"
-                           ToolTipService.Placement="Bottom" />
-            </MenuItem.Header>
-            <MenuItem StaysOpenOnClick="True">
-                <MenuItem.Header>
-                    <Calendar Name="monthCalendar"
-                              Style="{StaticResource CairoCalendarStyle}"
-                              HorizontalAlignment="Center"
-                              VerticalAlignment="Center" />
-                </MenuItem.Header>
-                <MenuItem.Template>
-                    <ControlTemplate TargetType="{x:Type MenuItem}">
-                        <ContentPresenter x:Name="HeaderHost"
-                                          RecognizesAccessKey="True"
-                                          ContentSource="Header" />
-                    </ControlTemplate>
-                </MenuItem.Template>
-            </MenuItem>
-            <MenuItem Click="OpenTimeDateCPL"
-                      Header="{Binding Path=(l10n:DisplayString.sMenuBar_OpenDateTimeSettings)}"
-                      Style="{StaticResource {x:Type MenuItem}}">
-                <MenuItem.Icon>
-                    <Image Source="{DynamicResource DateTimeIcon}"
-                           Width="16"
-                           Height="16" />
-                </MenuItem.Icon>
-            </MenuItem>
-        </MenuItem>
-    </Menu>
-</UserControl>
