--- Cairo Desktop/Extensions/CairoDesktop.Extensions.SystemMenuExtras/Search.xaml
@@ -1,82 +1,0 @@
-﻿<UserControl x:Class="CairoDesktop.Extensions.SystemMenuExtras.Search"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:controls="clr-namespace:CairoDesktop.Extensions.SystemMenuExtras"
-             xmlns:l10n="clr-namespace:CairoDesktop.Localization;assembly=CairoDesktop.Localization">
-    <Menu Style="{StaticResource CairoMenuBarMainContainerStyle}">
-        <MenuItem x:Name="CairoSearchMenu"
-                  Style="{StaticResource CairoMenuItemCairoSearchMenuStyle}"
-                  ItemContainerStyle="{StaticResource CairoMenuItemContainerStyle}"
-                  SubmenuOpened="FocusSearchBox">
-            <MenuItem.Header>
-                <Image Name="CairoSearchMenuIcon"
-                       Source="{DynamicResource SearchIcon}"
-                       Width="38"
-                       Height="22" />
-            </MenuItem.Header>
-            <StackPanel>
-                <StackPanel Orientation="Horizontal"
-                            FocusManager.FocusedElement="{Binding ElementName=searchStr}"
-                            Background="{DynamicResource ResourceKey=SearchTitleBackground}">
-                    <StackPanel Margin="15"
-                                Orientation="Horizontal">
-                        <TextBlock MinWidth="100"
-                                   Foreground="{DynamicResource ResourceKey=SearchTitleText}"
-                                   FontSize="{DynamicResource ResourceKey=LargeFontSize}"
-                                   FontFamily="{DynamicResource ResourceKey=GlobalFontFamily}"
-                                   Margin="0,1,0,0"
-                                   Text="{Binding Path=(l10n:DisplayString.sSearch_Title)}" />
-                        <Border Background="{DynamicResource ResourceKey=SearchResultBackground1}"
-                                CornerRadius="10"
-                                Margin="10,0,0,0"
-                                Padding="7,1,4,1">
-                            <StackPanel Orientation="Horizontal">
-                                <TextBox Style="{StaticResource SearchTextBox}"
-                                         Name="searchStr"
-                                         PreviewKeyDown="searchStr_KeyDown"
-                                         PreviewLostKeyboardFocus="searchStr_PreviewLostKeyboardFocus" />
-                                <Button Style="{StaticResource SearchTextBoxClearButton}"
-                                        Content="r"
-                                        Name="btnClearSearch"
-                                        Click="btnClearSearch_Click" />
-                            </StackPanel>
-                        </Border>
-                    </StackPanel>
-                </StackPanel>
-                <StackPanel Orientation="Vertical">
-                    <Button Style="{StaticResource CairoSearchResultsButtonStyle}"
-                            Name="btnViewResults"
-                            Click="btnViewResults_Click"
-                            Content="{Binding Path=(l10n:DisplayString.sSearch_ViewAllResults)}" />
-                    <ListView BorderThickness="0"
-                              Background="{x:Null}"
-                              MinHeight="0"
-                              MaxHeight="600"
-                              HorizontalAlignment="Stretch"
-                              Margin="-1"
-                              AlternationCount="2"
-                              Name="lstSearchResults"
-                              VirtualizingStackPanel.VirtualizationMode="Recycling"
-                              ScrollViewer.HorizontalScrollBarVisibility="Hidden">
-                        <ListView.ItemContainerStyle>
-                            <Style TargetType="ListViewItem"
-                                   BasedOn="{StaticResource CairoSearchResultsItemStyle}">
-                                <Setter Property="Template">
-                                    <Setter.Value>
-                                        <ControlTemplate TargetType="{x:Type ListViewItem}">
-                                            <Border BorderBrush="Transparent"
-                                                    BorderThickness="0"
-                                                    Background="{TemplateBinding Background}">
-                                                <controls:SearchResult DataContext="{Binding}" />
-                                            </Border>
-                                        </ControlTemplate>
-                                    </Setter.Value>
-                                </Setter>
-                            </Style>
-                        </ListView.ItemContainerStyle>
-                    </ListView>
-                </StackPanel>
-            </StackPanel>
-        </MenuItem>
-    </Menu>
-</UserControl>
