--- Cairo Desktop/Extensions/CairoDesktop.Extensions.SystemMenuExtras/SystemTray.xaml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<UserControl xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="d" xmlns:l10n="clr-namespace:CairoDesktop.Localization;assembly=CairoDesktop.Localization"
-             x:Class="CairoDesktop.Extensions.SystemMenuExtras.SystemTray"
-             x:Name="CairoSystemTray"
-             Height="22"
-             Margin="8,0,8,0">
-    <UserControl.Resources>
-        <Style TargetType="ItemsControl">
-            <Setter Property="Template">
-                <Setter.Value>
-                    <ControlTemplate TargetType="ItemsControl">
-                        <ItemsPresenter />
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-            <Setter Property="ItemsPanel">
-                <Setter.Value>
-                    <ItemsPanelTemplate>
-                        <StackPanel Orientation="Horizontal"
-                                    HorizontalAlignment="Right" />
-                    </ItemsPanelTemplate>
-                </Setter.Value>
-            </Setter>
-            <Setter Property="ItemTemplate">
-                <Setter.Value>
-                    <DataTemplate>
-                        <Border MouseUp="Image_MouseUp" MouseEnter="Image_MouseEnter" MouseLeave="Image_MouseLeave" MouseMove="Image_MouseMove" Background="Transparent" ToolTip="{Binding Path=Title}" ToolTipService.Placement="Bottom">
-                            <Image Source="{Binding Path=Icon, Mode=OneWay}"
-                                   Width="16"
-                                   Height="16"
-                                   Margin="5,0,5,2" />
-                        </Border>
-                    </DataTemplate>
-                </Setter.Value>
-            </Setter>
-        </Style>
-    </UserControl.Resources>
-    <StackPanel Orientation="Horizontal">
-        <ItemsControl ItemsSource="{Binding Path=UnpinnedIcons}" Name="UnpinnedItems" Visibility="Collapsed" />
-        <ToggleButton Name="btnToggle" Click="btnToggle_Click" Visibility="Collapsed" ToolTip="{Binding Path=(l10n:DisplayString.sMenuBar_ToggleNotificationArea)}" ToolTipService.Placement="Bottom" Style="{StaticResource CairoSystrayExpanderStyle}" />
-        <ItemsControl ItemsSource="{Binding Path=PinnedIcons}" Name="PinnedItems" />
-    </StackPanel>
-</UserControl>
